+++ CharacterData.h
@@ -53,7 +53,7 @@ namespace Evernus
 
         QString mName;
         QString mCorporationName;
-        uint mCorporationId = 0;
+        quint64 mCorporationId = 0;
         QString mRace;
         QString mBloodline;
         QString mAncestry;
